ecal-bridge/requirements.txt
./ecal5-5.10.3-1jammy-cp310-cp310-linux_x86_64.whl
python-socketio[client]
